requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.2
pillow==10.3.0
opencv-python==4.11.0.86
tensorflow==2.15.0
keras==2.15.0
streamlit==1.44.0
typing-extensions==4.7.1
protobuf==4.24.3
absl-py==1.4.0
flatbuffers==23.5.26
werkzeug==2.6.7
grpcio==1.57.0
h5py==3.9.0
six==1.17.1
python-dateutil==2.8.2
